refactor(everything): starting work from the ground up
